first version 4.25

+++ app/src/main/java/com/example/jareld/samsontransferclient/UserEvent.java
@@ -15,15 +15,15 @@ public class UserEvent {
     private long   progress;
     private String fileName;
 
-    public int getFileLengthMB() {
+    public long getFileLengthMB() {
         return fileLengthMB;
     }
 
-    public void setFileLengthMB(int fileLengthMB) {
+    public void setFileLengthMB(long fileLengthMB) {
         this.fileLengthMB = fileLengthMB;
     }
 
-    private int fileLengthMB;
+    private long fileLengthMB;
     public UserEvent(long l, String n) {
         name = n;
         progress = l;
